Make dbus tests be conditionally run based on pkg-config
The addition of the dbus tests requires dbus dev libraries be installed
to run the test suite. This is not always desirable or even possible.

So make building and running the dbus tests conditional on the
pkg-config info from those libs. If they are not present output a
message about skipping the tests.

This patch contains the review fix from sbeattie

Signed-off-by: John Johansen <john.johansen@canonical.com>
Acked-by: Steve Beattie <steve@nxnw.org>
Acked-by: Seth Arnold <seth.arnold@canonical.com>