Fix regular expressions in aa.py to enforce a space after some keywords (dbus, *mount, signal, ptrace, pivot_root) except if the line only contains the bare keyword.
Note that in most cases (except *mount) I used an alternation - this has 
the advantage that it doesn't change the match group numbering, with the
small disadvantage of having to mention the keyword twice in the regex.
I chose this way to avoid that I have to change lots of other places and
possibly introduce bugs by overlooking something.

For the *mount rules, I read the code - it shouldn't need any changes 
because it uses only matches[0..2] 

Acked-by: Tyler Hicks <tyhicks@canonical.com>